grouping common demos

+++ security/oidc/cleanup_oidc_demo.sh
@@ -1,10 +1,12 @@
 #!/usr/bin/env bash
 
+OIDC_PROVIDER=${OIDC_PROVIDER:-dex} # dex, google, sfdc
+
 # Get directory this script is located in to access script local files
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" &>/dev/null && pwd)"
 
-source "${SCRIPT_DIR}/../common_scripts.sh"
-source "${SCRIPT_DIR}/../working_environment.sh"
+source "${SCRIPT_DIR}/../../common_scripts.sh"
+source "${SCRIPT_DIR}/../../working_environment.sh"
 
 if [[ "${K8S_TOOL}" == 'kind' ]]; then
   KUBECONFIG=$(kind get kubeconfig-path --name="${DEMO_CLUSTER_NAME:-kind}")
@@ -12,7 +14,6 @@ if [[ "${K8S_TOOL}" == 'kind' ]]; then
 fi
 
 K8S_SECRET_NAME='my-oauth-secret'
-POLICY_K8S_CONFIGMAP='allow-jwt'
 
 cleanup_port_forward_deployment 'gateway-proxy-v2'
 
@@ -26,10 +27,8 @@ kubectl --namespace='default' delete \
   --filename="${GLOO_DEMO_RESOURCES_HOME}/petclinic-db.yaml" \
   --filename="${GLOO_DEMO_RESOURCES_HOME}/petclinic.yaml"
 
-cleanup_port_forward_deployment 'dex'
+if [[ "${OIDC_PROVIDER}" == 'dex' ]]; then
+  cleanup_port_forward_deployment 'dex'
 
-kubectl --namespace='gloo-system' delete \
-  --ignore-not-found='true' \
-  configmap/"${POLICY_K8S_CONFIGMAP}"
-
-helm delete --purge dex
+  helm delete --purge dex
+fi
